+++ src/bindings/nodejs/src/valhalla.cc
@@ -14,8 +14,8 @@
 #include <unordered_map>
 #include <utility>
 
-// Persistent reference to the RouterError JS constructor, initialized in Init()
-static Napi::FunctionReference RouterErrorConstructor;
+// Persistent reference to the ValhallaError JS constructor, initialized in Init()
+static Napi::FunctionReference ValhallaErrorConstructor;
 
 // Defined in graph_id.cc
 Napi::Object InitGraphId(Napi::Env env, Napi::Object exports);
@@ -83,7 +83,7 @@ protected:
       auto* actor = GetThreadLocalActor(config_);
       result_ = method_(actor, request_);
     } catch (const valhalla::valhalla_exception_t& e) {
-      // Store structured fields for OnError() to build a RouterError
+      // Store structured fields for OnError() to build a ValhallaError
       exception_code_ = e.code;
       exception_message_ = e.message;
       exception_http_code_ = e.http_code;
@@ -106,13 +106,13 @@ protected:
   }
 
   void OnError(const Napi::Error& e) override {
-    // If this was a valhalla_exception_t, reject with a RouterError instance
-    if (exception_code_ > 0 && !RouterErrorConstructor.IsEmpty()) {
+    // If this was a valhalla_exception_t, reject with a ValhallaError instance
+    if (exception_code_ > 0 && !ValhallaErrorConstructor.IsEmpty()) {
       Napi::Env env = Env();
-      auto err = RouterErrorConstructor.New({Napi::String::New(env, exception_message_),
-                                             Napi::Number::New(env, exception_code_),
-                                             Napi::Number::New(env, exception_http_code_),
-                                             Napi::String::New(env, exception_http_message_)});
+      auto err = ValhallaErrorConstructor.New({Napi::String::New(env, exception_message_),
+                                               Napi::Number::New(env, exception_code_),
+                                               Napi::Number::New(env, exception_http_code_),
+                                               Napi::String::New(env, exception_http_message_)});
       deferred_.Reject(err);
     } else {
       deferred_.Reject(e.Value());
@@ -318,8 +318,8 @@ Napi::Object Init(Napi::Env env, Napi::Object exports) {
   exports.Set(Napi::String::New(env, "VALHALLA_VERSION"),
               Napi::String::New(env, VALHALLA_PRINT_VERSION));
 
-  // Define RouterError as a JS class extending Error.
-  // Constructor signature: new RouterError(message, code, httpCode, httpMessage)
+  // Define ValhallaError as a JS class extending Error.
+  // Constructor signature: new ValhallaError(message, code, httpCode, httpMessage)
   auto routerErrorClass = Napi::Function::New(
       env,
       [](const Napi::CallbackInfo& info) {
@@ -339,9 +339,9 @@ Napi::Object Init(Napi::Env env, Napi::Object exports) {
         auto errorObj = Napi::Error::New(env, message);
         self.Set("stack", errorObj.Get("stack"));
       },
-      "RouterError");
+      "ValhallaError");
 
-  // Set up prototype chain: RouterError.prototype = Object.create(Error.prototype)
+  // Set up prototype chain: ValhallaError.prototype = Object.create(Error.prototype)
   auto global = env.Global();
   auto errorCtor = global.Get("Error").As<Napi::Function>();
   auto objectCtor = global.Get("Object").As<Napi::Function>();
@@ -350,13 +350,13 @@ Napi::Object Init(Napi::Env env, Napi::Object exports) {
 
   auto routerErrorProto = objectCreate.Call(objectCtor, {errorProto}).As<Napi::Object>();
   routerErrorProto.Set("constructor", routerErrorClass);
-  routerErrorProto.Set("name", Napi::String::New(env, "RouterError"));
+  routerErrorProto.Set("name", Napi::String::New(env, "ValhallaError"));
   routerErrorClass.Set("prototype", routerErrorProto);
 
-  RouterErrorConstructor = Napi::Persistent(routerErrorClass);
-  RouterErrorConstructor.SuppressDestruct();
+  ValhallaErrorConstructor = Napi::Persistent(routerErrorClass);
+  ValhallaErrorConstructor.SuppressDestruct();
 
-  exports.Set("RouterError", routerErrorClass);
+  exports.Set("ValhallaError", routerErrorClass);
 
   Actor::Init(env, exports);
   InitGraphId(env, exports);